Add script to run molecule only for changed roles
We add a script to run molecule scenario only on changed roles (based on
git history between two reference). This script is agnostic to complex
CI configuration (currently gitlab-ci) and can run localy.

We use Path instead of simple string, because it's more robust to handle
name with some char (Eg: /, \\,...) and also because we will allow users
to define parents with dedicated flags (pathlib.Path will be help to check
those inputs).