Add argument parser to allow user to define path
We add a way for user to define which directories our script has to
check.

For that we define a argparser with a subparser called test which can be
used whith the --parent flag(s). This change will be useful for people
who do not store molecule scenario and roles in standard collection
path.